rename Greenlet.link()'s argument to 'callback'
Previously it was called 'receiver'. Also link_exception() had
a default value for it which was None, but it did not work,
since linking to current greenlet feature was removed some time ago.

Section on spawn_link was removed from doc since these functions have
been gone for some time as well.

Closes #244.